LayoutTests/accessibility/mac/element-for-text-marker-expected.txt

text here
text block
text here
This tests the text marker system will return the correct element when given a text marker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.textMarkerRangeLength(textMarkerRange1) is 10
PASS text.accessibilityElementForTextMarker(startMarker).isEqual(text) is true
PASS text.accessibilityElementForTextMarker(endMarker).isEqual(text) is true
PASS text.textMarkerRangeLength(newMarkerRange) is 10
PASS newMarkerRange.isEqual(textMarkerRange1) is true
PASS successfullyParsed is true

TEST COMPLETE
